+++ packages/gatsby/src/commands/serve.ts
@@ -10,7 +10,6 @@ import report from "gatsby-cli/lib/reporter"
 import multer from "multer"
 import pathToRegexp from "path-to-regexp"
 import cookie from "cookie"
-
 import telemetry from "gatsby-telemetry"
 
 import { detectPortInUseAndPrompt } from "../utils/detect-port-in-use-and-prompt"
@@ -18,10 +17,19 @@ import { getConfigFile } from "../bootstrap/get-config-file"
 import { preferDefault } from "../bootstrap/prefer-default"
 import { IProgram } from "./types"
 import { IPreparedUrls, prepareUrls } from "../utils/prepare-urls"
+import { IGatsbyFunction } from "../redux/types"
 
 interface IMatchPath {
   path: string
   matchPath: string
+}
+
+interface IPathToRegexpKey {
+  name: string | number
+  prefix: string
+  suffix: string
+  pattern: string
+  modifier: string
 }
 
 interface IServeProgram extends IProgram {
@@ -120,10 +128,10 @@ module.exports = async (program: IServeProgram): Promise<void> => {
     `functions`
   )
 
-  let functions
+  let functions: Array<IGatsbyFunction> = []
   try {
     functions = JSON.parse(
-      fs.readFileSync(path.join(compiledFunctionsDir, `manifest.json`))
+      fs.readFileSync(path.join(compiledFunctionsDir, `manifest.json`), `utf-8`)
     )
   } catch (e) {
     // ignore
@@ -132,9 +140,9 @@ module.exports = async (program: IServeProgram): Promise<void> => {
   if (functions) {
     app.use(
       `/api/*`,
-      multer().none(),
+      multer().any(),
       express.urlencoded({ extended: true }),
-      (req, res, next) => {
+      (req, _, next) => {
         const cookies = req.headers.cookie
 
         if (!cookies) {
@@ -161,12 +169,13 @@ module.exports = async (program: IServeProgram): Promise<void> => {
           // We loop until we find the first match.
           functions.some(f => {
             let exp
-            const keys = []
+            const keys: Array<IPathToRegexpKey> = []
             if (f.matchPath) {
               exp = pathToRegexp(f.matchPath, keys)
             }
             if (exp && exp.exec(pathFragment) !== null) {
               functionObj = f
+              // @ts-ignore - TS bug? https://stackoverflow.com/questions/50234481/typescript-2-8-3-type-must-have-a-symbol-iterator-method-that-returns-an-iterato
               const matches = [...pathFragment.match(exp)].slice(1)
               const newParams = {}
               matches.forEach(