+++ packages/gatsby/src/redux/reducers/components.ts
@@ -22,6 +22,7 @@ export const componentsReducer = (
       if (!component) {
         component = {
           componentPath: action.payload.componentPath,
+          componentChunkName: action.payload.componentChunkName,
           query: ``,
           pages: new Set(),
           isInBootstrap: true,