+++ packages/gatsby/src/utils/flags.ts
@@ -176,6 +176,30 @@ const activeFlags: Array<IFlag> = [
     umbrellaIssue: `https://gatsby.dev/functions-feedback`,
     testFitness: (): fitnessEnum => true,
   },
+  {
+    name: `CONCURRENT_FEATURES`,
+    env: `GATSBY_EXPERIMENTAL_CONCURRENT_FEATURES`,
+    command: `all`,
+    telemetryId: `ConcurrentFeatures`,
+    experimental: true,
+    description: `Enable React's concurrent features`,
+    // umbrellaIssue: `https://gatsby.dev/concurrent-features`,
+    testFitness: (): fitnessEnum => {
+      // Because of this, this flag will never show up
+      const semverConstraints = {
+        react: `^0.0.0-experimental-57768ef90`,
+        "react-dom": `^0.0.0-experimental-57768ef90`,
+      }
+
+      if (satisfiesSemvers(semverConstraints)) {
+        return true
+      } else {
+        // react & react-dom is either not installed or not new enough so
+        // just disable — it won't work anyways.
+        return false
+      }
+    },
+  },
 ]
 
 export default activeFlags